refactor(security): unify Signer and Codec into FeedToken
Consolidates the FeedToken value object definition alongside its nested Codec and Signer namespaces inside feed_token.rb, removing the subfolder structure and keeping token verification, compression, and signature validation cohesive.
